Updated avatar color and view all link
